feat: Allow unified responses for all feedback datasets
